% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwoCpp}
\alias{timesTwoCpp}
\title{Multiply by two (Rcpp version)}
\usage{
timesTwoCpp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector that is double the input vector
}
\description{
This function multiples numbers by two, using C++ code.
}
