% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesTwoR.R
\name{timesTwoR}
\alias{timesTwoR}
\title{Multiply by two (R version).}
\usage{
timesTwoR(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A numeric vector that is double the input vector
}
\description{
This function multiplies numbers by two.
}
\examples{
timesTwoR(2)
timesTwoR(1:3)

}
